tips front

+++ src/model/tips.requests.php
@@ -14,6 +14,9 @@ function getTips()
 
 function addTips($content)
 {
+    if (!userIsAdmin()) {
+        return false;
+    }
     global $db;
     $query = $db->prepare('INSERT INTO tips (content) VALUES (:content)');
     $query->execute([
@@ -23,9 +26,22 @@ function addTips($content)
 
 function deleteTips($id)
 {
+    if (!userIsAdmin()) {
+        return false;
+    }
     global $db;
     $query = $db->prepare('DELETE FROM tips WHERE id = :id');
     $query->execute([
         'id' => $id,
     ]);
+}
+
+function getTipsById($id)
+{
+    global $db;
+    $query = $db->prepare('SELECT * FROM tips WHERE id = :id');
+    $query->execute([
+        'id' => $id,
+    ]);
+    return $query->fetch();
 }
